requirements.txt
langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.3.0
langchain-openai>=1.0.0
langgraph>=1.0.0
langgraph-checkpoint>=3.0.0
dashscope>=1.24.0
pyyaml==6.0.3
rich==14.2.0
flask==3.0.0
flask-cors==4.0.0
langchain-tavily
mcp>=1.0.0
chromadb>=0.4.0
